requirements.txt
beautifulsoup4==4.9.1
google-api-python-client==1.8.0
lxml==4.5.2
mutagen==1.45.1
Pillow==7.2.0
pygame==1.9.6
pyinstaller==4.0
requests==2.24.0
youtube-dl==2020.7.28